refactor(#450): move common parser to deck-utils
