+++ remote/src/app/utils/parse-slides.utils.tsx
@@ -4,9 +4,10 @@ import uuid from 'uuid/v4';
 
 import {DeckdeckgoDeckDefinition, DeckdeckgoSlideDefinition} from '@deckdeckgo/types';
 
+import {cleanContent} from '@deckdeckgo/deck-utils';
+
 import {ParseElementsUtils} from './parse-elements.utils';
 import {ParseAttributesUtils} from './parse-attributes.utils';
-import {ParseContentUtils} from './parse-content.utils';
 
 export class ParseSlidesUtils {
 
@@ -46,7 +47,7 @@ export class ParseSlidesUtils {
 
             // Create a div to parse back to JSX its children
             const div = document.createElement('div');
-            div.innerHTML = await ParseContentUtils.cleanContent(slide.content);
+            div.innerHTML = await cleanContent(slide.content);
 
             const content = await ParseElementsUtils.parseElements(div, true);
