+++ studio/src/app/pages/editor/app-editor/app-editor.tsx
@@ -7,12 +7,13 @@ import {filter, take} from 'rxjs/operators';
 
 import {isFullscreen, isMobile, debounce, isIOS} from '@deckdeckgo/utils';
 
+import {convertStyle} from '@deckdeckgo/deck-utils';
+
 import {AuthUser} from '../../../models/auth/auth.user';
 import {SlideTemplate} from '../../../models/data/slide';
 import {Deck} from '../../../models/data/deck';
 
 import {CreateSlidesUtils} from '../../../utils/editor/create-slides.utils';
-import {ParseStyleUtils} from '../../../utils/editor/parse-style.utils';
 import {ParseBackgroundUtils} from '../../../utils/editor/parse-background.utils';
 import {IonControllerUtils} from '../../../utils/core/ion-controller-utils';
 
@@ -230,7 +231,7 @@ export class AppEditor {
         return new Promise<void>((resolve) => {
             this.deckEditorService.watch().pipe(take(1)).subscribe(async (deck: Deck) => {
                 if (deck && deck.data && deck.data.attributes && deck.data.attributes.style) {
-                    this.style = await ParseStyleUtils.convertStyle(deck.data.attributes.style);
+                    this.style = await convertStyle(deck.data.attributes.style);
                 } else {
                     this.style = undefined;
                 }
