+++ utils/deck/src/index.ts
@@ -1,0 +1,2 @@
+export * from './utils/deckdeckgo-parse-style.utils';
+export * from './utils/deckdeckgo-parse-content.utils';
